# Drop Cached Polynomials