@aszepieniec 2024-09-29
Comparing profiles for Fib 100 one finds almost identical timings except for the out-of-domain rows. On master:
```
├─out-of-domain rows                            2.88s
```
on this branch:
```
├─out-of-domain rows                           36.50s
```

The discrepancy comes from having to use barycentric evaluation instead of Horner evaluation (since we do not have the polynomials anymore). If barycentric evaluation cannot be made any faster, perhaps we should consider cashing the polynomials after all for memory-inefficient code path.