@Sword-Smith 2024-09-29
IIRC our current (host machine) Horner evaluation uses memory allocation. In the PR that introduces (host machine) Horner evaluation, I proposed an alternative version that doesn't allocate. It's worth trying that one to see if it solves the problem.