@Sword-Smith 2024-09-30
> IIRC our current (host machine) barycentric evaluation uses memory allocation. In the PR that introduces (host machine) barycentric evaluation, I proposed an alternative version that doesn't allocate. It's worth trying that one to see if it solves the problem.

Try implementing barycentric evaluation like [this](https://github.com/TritonVM/triton-vm/pull/266#issuecomment-2072131685) instead. It avoids the allocation and might give a meaningful speedup.