@Sword-Smith 2024-09-30
> (...) Some benchmarks, obtain on my laptop. (...)

On mjolnir, only measuring the "out-of-domain rows" step:

| fib parameter | version | time |
|---|---|---|---|
|40 000 | horner | 430.98ms  |
| 40 000 | batch barycentric | 580.05ms|

So even though we see a small slowdown, I think it's worth it to save the RAM and drop the cached interpolants.