@Sword-Smith 2024-09-30
**Proving with full caching**

This PR:
```
Prove Fibonacci 40000   time:   [29.757 s 29.974 s 30.182 s]     
```

`master`
```
Prove Fibonacci 40000   time:   [29.104 s 29.216 s 29.317 s]                  
                                           change: [-3.3020% -2.5289% -1.7164%] (p = 0.00 < 0.05)
```

So this PR introduces a 2.5 % slowdown on the fully cached code path.